Made waiting more alert for enemies

+++ src/gameplay/resources/player_action_state.rs
@@ -88,7 +88,7 @@ impl PlayerActionState {
                     None // process the cancellation next turn
                 } else {
                     Some(Action::Stay {
-                        duration: StayDuration::Long,
+                        duration: StayDuration::Short,
                     })
                 }
             }
